requirements.txt
torch
numpy
tiktoken
datasets
tqdm
matplotlib
tensorboard
rich
h5py
wandb
pydantic
PyYAML